Move shortcuts code to other file, change LiveViewManager initialization to improve lazy initialization support
